requirements.txt
Flask==2.3.3
yt-dlp==2024.5.27
instagrapi==1.20.6
APScheduler==3.10.4
moviepy==1.0.3
pillow==10.3.0
python-dotenv==1.0.1
requests==2.32.2
